STAC-12178 Fixes backup/restore issues in Helm chart discovered while migrating preprod instance of StackState

See merge request stackvista/devops/helm-charts!399

+++ stable/stackstate/restore/restore-stackgraph-backup.sh
@@ -8,8 +8,21 @@ JOB_NAME=stackgraph-restore-$(date +%Y%m%dt%H%M%S)
 JOB_YAML_DIR=$(mktemp -d /tmp/sts-restore-XXXXXX)
 JOB_YAML_FILE="${JOB_YAML_DIR}/job-${JOB_NAME}.yaml"
 
-kubectl get configmap stackstate-backup-restore-scripts -o jsonpath="{.data.job-${JOB_NAME_TEMPLATE}\.yaml}"  | sed -e "s/${JOB_NAME_TEMPLATE}/${JOB_NAME}/" -e "s/REPLACE_ME_BACKUP_FILE_REPLACE_ME/${BACKUP_FILE}/" > "${JOB_YAML_FILE}"
+if (! (kubectl get configmap stackstate-backup-restore-scripts -o jsonpath="{.data.job-${JOB_NAME_TEMPLATE}\.yaml}"  | sed -e "s/${JOB_NAME_TEMPLATE}/${JOB_NAME}/" -e "s/REPLACE_ME_BACKUP_FILE_REPLACE_ME/${BACKUP_FILE}/" > "${JOB_YAML_FILE}")) || [ ! -s "${JOB_YAML_FILE}" ]; then
+    echo "Did you set backup.stackGraph.restore.enabled to true?"
+    exit 1
+fi
+
+echo "=== Scaling down deployments for pods that connect to StackGraph"
+kubectl scale --replicas=0 deployments --selector=stackstate.com/connects-to-stackgraph=true
+
+echo "=== Allowing pods to terminate"
+sleep 15
+
+echo "=== Starting restore job"
 kubectl create -f "${JOB_YAML_FILE}"
-while ! kubectl logs job/"${JOB_NAME}" --container=restore >/dev/null 2>/dev/null ; do echo "Waiting for job to start..."; sleep 2; done; kubectl logs "job/${JOB_NAME}"  --container=restore --follow=true
-kubectl delete job "${JOB_NAME}"
 rm -rf "${JOB_YAML_DIR}"
+
+echo "=== Restore job started. Follow the logs and clean up the job with the following commands:"
+echo "kubectl logs --follow job/${JOB_NAME}"
+echo "kubectl delete job/${JOB_NAME}"
